refactor: split ops.test.js into separate test files by concern
Split monolithic test file into tensor, nn, autograd, module, and native
test files. Add ESM loader for .js->.ts resolution. Remove obsolete
toy-specific test stubs (sin/cos/sqrt) that don't belong on main.
